Adiciona testes de integração

+++ src/main/kotlin/br/com/mcos/pockotlinspringboot/controller/dto/NovoJogoRequest.kt
@@ -1,10 +1,11 @@
 package br.com.mcos.pockotlinspringboot.controller.dto
 
 import br.com.mcos.pockotlinspringboot.domain.Jogo
+import java.math.BigDecimal
 
 data class NovoJogoRequest (
     val titulo: String,
-    val preco: Int,
+    val preco: BigDecimal,
     val ano: Int
 ){
     fun paraJogo() = Jogo(
